cleaned up width and created separate file for class
