+++ Braze Demo/View/Grouped/SectionBackgroundDecorationView.swift
@@ -1,0 +1,13 @@
+import UIKit
+
+class SectionBackgroundDecorationView: UICollectionReusableView {
+  override init(frame: CGRect) {
+    super.init(frame: frame)
+    
+    backgroundColor = .systemGroupedBackground
+  }
+  
+  required init?(coder: NSCoder) {
+    super.init(coder: coder)
+  }
+}
